feat(fake): refactor fake mission to be able to reuse it
 - create a fake action  to have data
 - Update current files using fake daata
