+++ backend/src/main/kotlin/fr/gouv/dgampa/rapportnav/domain/use_cases/mission/FakeMissionData.kt
@@ -7,7 +7,9 @@ import fr.gouv.dgampa.rapportnav.domain.entities.mission.env.ActionCompletionEnu
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.env.MissionSourceEnum
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.env.MissionTypeEnum
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.env.envActions.*
+import fr.gouv.dgampa.rapportnav.domain.entities.mission.fish.fishActions.MissionActionType
 import fr.gouv.dgampa.rapportnav.domain.entities.mission.nav.action.ExtendedEnvActionEntity
+import fr.gouv.dgampa.rapportnav.domain.entities.mission.nav.action.ExtendedFishActionEntity
 import fr.gouv.dgampa.rapportnav.domain.entities.user.User
 import org.locationtech.jts.geom.Coordinate
 import java.time.Instant
@@ -15,7 +17,7 @@ import java.util.*
 
 @UseCase
 class FakeMissionData(
-    private val getFishActionsByMissionId: GetFishActionsByMissionId,
+    private val getFakeActionData: FakeActionData
 ) {
 
     fun getEmptyMissionIds(): List<Int> {
@@ -103,9 +105,7 @@ class FakeMissionData(
         val envActions = envControls(missionId).map {
             MissionActionEntity.EnvAction(it)
         }
-        val fishActions = getFishActionsByMissionId.getFakeActions(missionId).map {
-            MissionActionEntity.FishAction(it)
-        }
+        val fishActions = getFishActions(missionId)
 
         return MissionEntity(
             id = missionId,
@@ -120,6 +120,19 @@ class FakeMissionData(
             openBy = "fake",
             actions = fishActions + envActions
         )
+    }
+
+    fun getFishActions(missionId: Int) = getFakeActionData.getFakeFishActions(missionId).filter {
+        listOf(
+            MissionActionType.SEA_CONTROL,
+            MissionActionType.LAND_CONTROL
+        ).contains(it.actionType)
+    }.map {
+        var action = ExtendedFishActionEntity.fromMissionAction(it)
+        action.controlAction?.computeCompleteness()
+        action
+    }.map {
+        MissionActionEntity.FishAction(it)
     }
 
     fun emptyMission(missionId: Int): MissionEntity {
